# backend/src/openapi_server/models/craftsman.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Craftsman(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Craftsman - a model defined in OpenAPI

        id: The id of this Craftsman [Optional].
        name: The name of this Craftsman [Optional].
        ranking_score: The ranking_score of this Craftsman [Optional].
    """

    id: Optional[int] = Field(alias="id", default=None)
    name: Optional[str] = Field(alias="name", default=None)
    ranking_score: Optional[float] = Field(alias="rankingScore", default=None)

Craftsman.update_forward_refs()