# backend/src/openapi_server/models/patch_request.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class PatchRequest(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    PatchRequest - a model defined in OpenAPI

        max_driving_distance: The max_driving_distance of this PatchRequest [Optional].
        profile_picture_score: The profile_picture_score of this PatchRequest [Optional].
        profile_description_score: The profile_description_score of this PatchRequest [Optional].
    """

    max_driving_distance: Optional[float] = Field(alias="maxDrivingDistance", default=None)
    profile_picture_score: Optional[float] = Field(alias="profilePictureScore", default=None)
    profile_description_score: Optional[float] = Field(alias="profileDescriptionScore", default=None)

PatchRequest.update_forward_refs()