# backend/src/openapi_server/models/patch_response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.patch_response_updated import PatchResponseUpdated


class PatchResponse(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    PatchResponse - a model defined in OpenAPI

        id: The id of this PatchResponse [Optional].
        updated: The updated of this PatchResponse [Optional].
    """

    id: Optional[int] = Field(alias="id", default=None)
    updated: Optional[PatchResponseUpdated] = Field(alias="updated", default=None)

PatchResponse.update_forward_refs()