# backend/src/openapi_server/models/response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.craftsman import Craftsman


class Response(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Response - a model defined in OpenAPI

        craftsmen: The craftsmen of this Response [Optional].
    """

    craftsmen: Optional[List[Craftsman]] = Field(alias="craftsmen", default=None)

Response.update_forward_refs()